Implement q_remove_head() and q_remove_tail()
Implement functions that remove (but not delete, the space used by the
list element and the string is not freed) an element from head and tail
of a queue. If the supplied string is non-NULL, then the removed string
is copied to *sp. If the queue is NULL or empty, it does nothing and
return NULL.